+++ tla-bmcmt/src/test/scala/at/forsyte/apalache/tla/bmcmt/stratifiedRules/BoolRewriterTest.scala
@@ -1,6 +1,5 @@
 package at.forsyte.apalache.tla.bmcmt.stratifiedRules
 
-import at.forsyte.apalache.tla.bmcmt.stratifiedRules.aux.RewriterScope
 import at.forsyte.apalache.tla.bmcmt.{ArenaCell, Binding, PureArena}
 import at.forsyte.apalache.tla.lir._
 import at.forsyte.apalache.tla.types.tla
