+++ tla-bmcmt/src/test/scala/at/forsyte/apalache/tla/bmcmt/stratifiedRules/FunRewriterTest.scala
@@ -1,9 +1,5 @@
 package at.forsyte.apalache.tla.bmcmt.stratifiedRules
 
-import at.forsyte.apalache.tla.bmcmt.{ArenaCell, Binding, PureArena}
-import at.forsyte.apalache.tla.bmcmt.stratifiedRules.aux.RewriterScope
-import at.forsyte.apalache.tla.lir._
-import at.forsyte.apalache.tla.types.tla
 import org.junit.runner.RunWith
 import org.scalatest.BeforeAndAfterEach
 import org.scalatest.funsuite.AnyFunSuite
@@ -35,7 +31,7 @@ class FunRewriterTest extends AnyFunSuite with BeforeAndAfterEach {
 //    val startScope = RewriterScope(arena, Binding(Map("x" -> xCell, "y" -> yCell)))
 //
 //    val (endScope, cell) = rewriter.rewrite(ex)(startScope)
-//s
+//
   }
 
 }
